Refactor generateGrid() to properly create grid
Using 2d array, creating the grid is possible. For each row of the array will be assigned to a parent row container div which will contain its children. The children are to be assigned in the inner for loop which is the columns in the array.

After each column is created and fills a row, we then append it to the parent container to render it.